Backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==4.2
django-allauth==0.63.3
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-spectacular==0.27.2
flake8==7.1.0
idna==3.8
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
Pillow==9.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
sorl-thumbnail==12.10.0
sqlparse==0.5.1
typing-extensions==4.12.2
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.2
